Conflicts:
	Charts/vtkScatterPlotMatrix.cxx
	Filtering/vtkImageData.cxx
	Filtering/vtkImageData.h
	Hybrid/vtkAxisActor.cxx
	Hybrid/vtkLSDynaReader.cxx
	IO/vtkImageWriter.cxx
	IO/vtkPLOT3DReader.cxx
	IO/vtkSESAMEReader.cxx
	Infovis/vtkDelimitedTextReader.cxx
	Parallel/vtkPImageWriter.cxx
	Views/vtkDataRepresentation.cxx
	VolumeRendering/vtkVolumeOutlineSource.cxx
	Widgets/vtkPointPlacer.cxx

+++ Utilities/vtkfreetype/src/base/ftbitmap.c
@@ -418,7 +418,7 @@
         target->pitch = source->width + pad;
 
         if ( target->pitch > 0                           &&
-             target->rows > FT_ULONG_MAX / target->pitch )
+             target->rows > (int)(FT_ULONG_MAX / target->pitch) )
           return FT_Err_Invalid_Argument;
 
         if ( target->rows * target->pitch > old_size             &&
